move ByteArray to //common/collection from //common/util
cleanup

fix flipped boolean

caching hash codes is 10% faster
