+++ concept/thing/impl/ThingImpl.java
@@ -18,10 +18,9 @@
 
 package com.vaticle.typedb.core.concept.thing.impl;
 
-import com.vaticle.typedb.common.collection.Bytes;
 import com.vaticle.typedb.core.common.exception.TypeDBException;
 import com.vaticle.typedb.core.common.iterator.FunctionalIterator;
-import com.vaticle.typedb.core.common.util.ByteArray;
+import com.vaticle.typedb.core.common.collection.ByteArray;
 import com.vaticle.typedb.core.concept.Concept;
 import com.vaticle.typedb.core.concept.ConceptImpl;
 import com.vaticle.typedb.core.concept.thing.Attribute;
