+++ graph/iid/IID.java
@@ -18,7 +18,7 @@
 
 package com.vaticle.typedb.core.graph.iid;
 
-import com.vaticle.typedb.core.common.util.ByteArray;
+import com.vaticle.typedb.core.common.collection.ByteArray;
 
 public abstract class IID {
