+++ graph/vertex/impl/TypeVertexImpl.java
@@ -20,7 +20,7 @@ package com.vaticle.typedb.core.graph.vertex.impl;
 
 import com.vaticle.typedb.core.common.iterator.FunctionalIterator;
 import com.vaticle.typedb.core.common.parameters.Label;
-import com.vaticle.typedb.core.common.util.ByteArray;
+import com.vaticle.typedb.core.common.collection.ByteArray;
 import com.vaticle.typedb.core.graph.TypeGraph;
 import com.vaticle.typedb.core.graph.adjacency.TypeAdjacency;
 import com.vaticle.typedb.core.graph.adjacency.impl.TypeAdjacencyImpl;
@@ -36,7 +36,7 @@ import java.util.function.Supplier;
 import java.util.regex.Pattern;
 
 import static com.vaticle.typedb.core.common.iterator.Iterators.link;
-import static com.vaticle.typedb.core.common.util.ByteArray.join;
+import static com.vaticle.typedb.core.common.collection.ByteArray.join;
 import static com.vaticle.typedb.core.graph.common.Encoding.Edge.Type.OWNS;
 import static com.vaticle.typedb.core.graph.common.Encoding.Edge.Type.OWNS_KEY;
 import static com.vaticle.typedb.core.graph.common.Encoding.Edge.Type.PLAYS;
@@ -355,7 +355,7 @@ public abstract class TypeVertexImpl extends VertexImpl<VertexIID.Type> implemen
         }
 
         private void commitPropertyScope() {
-            graph.storage().putUntracked(join(iid.bytes(), SCOPE.infix().bytes()), ByteArray.of(scope.getBytes()));
+            graph.storage().putUntracked(join(iid.bytes(), SCOPE.infix().bytes()), ByteArray.encodeString(scope));
         }
 
         private void commitPropertyAbstract() {
@@ -363,7 +363,7 @@ public abstract class TypeVertexImpl extends VertexImpl<VertexIID.Type> implemen
         }
 
         private void commitPropertyLabel() {
-            graph.storage().putUntracked(join(iid.bytes(), LABEL.infix().bytes()), ByteArray.of(label.getBytes()));
+            graph.storage().putUntracked(join(iid.bytes(), LABEL.infix().bytes()), ByteArray.encodeString(label));
         }
 
         private void commitPropertyValueType() {
@@ -371,7 +371,7 @@ public abstract class TypeVertexImpl extends VertexImpl<VertexIID.Type> implemen
         }
 
         private void commitPropertyRegex() {
-            graph.storage().putUntracked(join(iid.bytes(), REGEX.infix().bytes()), ByteArray.of(regex.pattern().getBytes()));
+            graph.storage().putUntracked(join(iid.bytes(), REGEX.infix().bytes()), ByteArray.encodeString(regex.pattern()));
         }
     }
 
@@ -411,7 +411,7 @@ public abstract class TypeVertexImpl extends VertexImpl<VertexIID.Type> implemen
         public void label(String label) {
             assert !isDeleted();
             graph.update(this, this.label, scope, label, scope);
-            graph.storage().putUntracked(join(iid.bytes(), LABEL.infix().bytes()), ByteArray.of(label.getBytes()));
+            graph.storage().putUntracked(join(iid.bytes(), LABEL.infix().bytes()), ByteArray.encodeString(label));
             graph.storage().deleteUntracked(IndexIID.Type.Label.of(this.label, scope).bytes());
             graph.storage().putUntracked(IndexIID.Type.Label.of(label, scope).bytes(), iid.bytes());
             this.label = label;
@@ -421,7 +421,7 @@ public abstract class TypeVertexImpl extends VertexImpl<VertexIID.Type> implemen
         public void scope(String scope) {
             assert !isDeleted();
             graph.update(this, label, this.scope, label, scope);
-            graph.storage().putUntracked(join(iid.bytes(), SCOPE.infix().bytes()), ByteArray.of(scope.getBytes()));
+            graph.storage().putUntracked(join(iid.bytes(), SCOPE.infix().bytes()), ByteArray.encodeString(scope));
             graph.storage().deleteUntracked(IndexIID.Type.Label.of(label, this.scope).bytes());
             graph.storage().putUntracked(IndexIID.Type.Label.of(label, scope).bytes(), iid.bytes());
             this.scope = scope;
@@ -476,7 +476,7 @@ public abstract class TypeVertexImpl extends VertexImpl<VertexIID.Type> implemen
             assert !isDeleted();
             if (regex == null) graph.storage().deleteUntracked(join(iid.bytes(), REGEX.infix().bytes()));
             else
-                graph.storage().putUntracked(join(iid.bytes(), REGEX.infix().bytes()), ByteArray.of(regex.pattern().getBytes()));
+                graph.storage().putUntracked(join(iid.bytes(), REGEX.infix().bytes()), ByteArray.encodeString(regex.pattern()));
             this.regex = regex;
             this.setModified();
             return this;
